Relaunch release in-dev Make all request must apply Account object. Add helper method to some entity class for robust. Fix CreateTracker eager behavior. Add new ContextMenu command to parse Crayon Record. CrayonRecord add new field for input RecordDate. Redesign container / embed object that display to discord Rewiring some interaction for Public Posting to directly called. Remove unused method in Handler (for now) as well as SessionManager some session are not used already.

+++ global/src/main/java/io/github/zazalng/prickcal/global/contract/trickcal/crayon/CrayonCosts.java
@@ -18,16 +18,16 @@
 package io.github.zazalng.prickcal.global.contract.trickcal.crayon;
 
 public enum CrayonCosts {
-    HOUSE_1((short) 1, 2, 3),
-    HOUSE_2((short) 2, 4, 4),
-    HOUSE_3((short) 3, 6, 5),
-    UNKNOWN((short) 0, -1, -1);
+    HOUSE_1(1, 2, 3),
+    HOUSE_2(2, 4, 4),
+    HOUSE_3(3, 6, 5),
+    UNKNOWN(0, -1, -1);
 
-    private final short depth;
+    private final int depth;
     private final int price;
     private final int amount;
 
-    CrayonCosts(short depth, int price, int amount) {
+    CrayonCosts(int depth, int price, int amount) {
         this.depth = depth;
         this.price = price;
         this.amount = amount;
@@ -41,7 +41,7 @@ public enum CrayonCosts {
         return amount;
     }
 
-    public static CrayonCosts fromDepth(short depth) {
+    public static CrayonCosts fromDepth(int depth) {
         for(CrayonCosts e: CrayonCosts.values()){
             if (e.depth == depth) return e;
         }
